refactor(GameLog): change methods - 메서드 이름과 매개변수를 좀더 가독성 있게 변경했다. - 필요없는 기능을 통폐합했다.
