+++ src/main/java/racingcar/domain/receiver/TryCountReceiver.java
@@ -1,4 +1,4 @@
-package racingcar.receiver;
+package racingcar.domain.receiver;
 
 import racingcar.exception.TryCountException;
