+++ src/test/java/racingcar/receiver/CarNamesReceiverTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.receiver.CarNamesReceiver;
 import racingcar.exception.CarNameException;
 
 public class CarNamesReceiverTest {
