+++ src/test/java/racingcar/receiver/TryCountReceiverTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.receiver.TryCountReceiver;
 import racingcar.exception.TryCountException;
 
 public class TryCountReceiverTest {
